% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{equator}
\alias{equator}
\title{Celestial equator is the great circle with declination equal to 0 deg.}
\format{
data.frame
\describe{
  \item{l}{Galactic longitude, degrees.}
  \item{b}{Galactic latitude, degrees.}
  \item{ra}{Right ascension, degrees.}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Celestial_equator}
}
\usage{
equator
}
\description{
Celestial equator is the great circle with declination equal to 0 deg.
}
\keyword{datasets}
