% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_equatorial.R
\name{scale_eq_ra}
\alias{scale_eq_ra}
\title{Equatorial right-ascension scale settings}
\usage{
scale_eq_ra(breaks = ggplot2::waiver(), minor_breaks = ggplot2::waiver())
}
\arguments{
\item{breaks, minor_breaks}{Break specification passed to `coord_equatorial()`.}
}
\value{
An object consumed by `ggplot_add()`.
}
\description{
Equatorial right-ascension scale settings
}
