% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sky_scales.R
\name{scale_gal_lat}
\alias{scale_gal_lat}
\title{Galactic latitude scale settings}
\usage{
scale_gal_lat(breaks = ggplot2::waiver(), minor_breaks = ggplot2::waiver())
}
\arguments{
\item{breaks, minor_breaks}{Break specification passed to `coord_galactic()`.}
}
\value{
An object consumed by `ggplot_add()`.
}
\description{
Galactic latitude scale settings
}
